requirements.txt
pyltp~=0.2.1
pandas~=1.1.5
Keras~=2.4.3
numpy~=1.19.5
tensorflow~=2.5.0
fasttext~=0.9.2
scikit-learn~=0.24.2
gensim~=4.1.2
Pillow~=8.3.1
pypinyin~=0.46.0
Pinyin2Hanzi~=0.1.1